[MOGURI-6] fix : goal CRUD_CREATE 및 UPDATE 수정 완료

+++ moguri/src/main/java/org/moguri/goal/domain/Goal.java
@@ -14,7 +14,7 @@ import java.util.Date;
 @Builder
 public class Goal {
     private long goalId;
-    private int memberId;
+    private long memberId;
     private String goalName;
     private BigDecimal goalAmount;
     private BigDecimal currentAmount;
